backend/pyproject.toml
[tool.poetry]
name = "school-management-backend"
version = "0.1.0"
description = "Backend for School Management System"
authors = ["OpenHands <openhands@all-hands.dev>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"
redis = ">=4.2.0rc1,<5.0.0"
celery = "^5.3.6"
flower = "^2.0.1"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
aiofiles = "^23.2.1"
boto3 = "^1.34.7"
python-dotenv = "^1.0.0"
prometheus-client = "^0.19.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.1"}
structlog = "^23.2.0"
slowapi = "^0.1.8"
secure = "^0.3.0"
tenacity = "^8.2.3"
httpx = "^0.25.2"
python-json-logger = "^2.0.7"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"
locust = "^2.19.1"
safety = "^2.3.5"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]